Special case when using EVL for the outer mask instead of nullptr
Representing the outer mask as null when logically, under EVL, is not
needed works fine most of the time, except when there are legit users of
the predicate itself. This happens during reductions that combine
reduced vector using a whole-register select.

Rather than representing the outer mask using nullptr, special case the
ICmpULE VPInstruction which exists only for that. This way, if there is
a legit user of the mask for other reasons than masking, we can still
provide one.