Do not use a class that is a singleton but hardcode V0
Any derived class we can built from it via intersection will be a
singleton as well. This restricts the register allocator a lot and
causes register allocation failures (the virtual register illusion
vanishes hard with singleton register classes).

Using V0 in the pattern means we will emit a copy. Often this copy
can be removed because it is a trivial $v0 = COPY $v0.