Disable vectorisation for real
Clang supports vectorisation and interleaving as two different
transformations implemented by the same pass. Both can be enabled or
disabled independently.

Unfortunately there is a way to disable the pass but not either of
transformations, so the current technique relies on setting the vector
factor or the interleave count to neutral values that have no effect.
The value is fixed(1).

This works fine for the interleave factor but is problematic for vector
factor, specially in our case where we do only scalable vectorisation:
turns out that when we disable vectorization in this case the vector
width is interpreted as scalable(1). So effectively vectorisation is not
possible to disable in this case and it runs as if we were vectorising
with a vector factor of scalable(1).

We are currently not using interleave at all (without vectorization) so
it makes sense to us that vectorize(disable) does disable the
vectorization rather than trying to make it a no-op, which turns
it is fragile when using scalable vectorisation.