Force use of x7 when updaring sp in the prologue/epilogue
When adjusting the fp or sp we may have to create a virtual register if
the offset requested won't fit a 5-bit immediate. The Register Scavenger
can deal with this virtual register. This usually works well but we have
seen that test SingleSource/UnitTests/matrix-types-spec.test of the
testsuite creates a humongous entry basic block.

This basic block is so large that it needs many fixes from the register
scavenger. The stack is so large as well that when we grow it at the
prolog we need a virtual register. It turns out RS doesn't find any
available and chooses a2. Then it realises a2 is live and it puts in the
spill slot (found at sp + 0). But a2 was spilled just to enlarge the
stack, rendering the earlier sp + 0 an invalid offset. Later on a2 is
reloaded and we keep using sp + 0 under the assumption the stack pointer
is not changing for the emergency slot. Our current support for the
emergency slot cannot work in this circumstance.

Something like this is emitted in the prologue:

  sd a2, 0(sp)
  lui a2, 4
  addiw a2, a2, -288
  sub sp, sp, a2      # oops!

Later on we do

  ld a2, 0(sp)        # oops!

So, as a stop-gap solution until we figure out a better approach here,
we teach the adjustment function to always use t2 (x7) as a scratch
register when updating the stack pointer. This is safe to do in the
prologue and the epilogue but not in the other places of the function
which need adjusting the stack pointer (e.g. in a function call) where
the RS should be able to find a register or successfully using the
emergency spill slot.