Move the calculation of EVL as early as possible
We may have to move some recipes around in the LV and some of
those recipes use EVL. Because we create it lazily it may end
after the use.

Ideally we should do something like we do for TripCount, so we
avoid having a recipe around that may end misplaced.

This fixes the crash in #326 but not the underlying mess
we have with phis.