Workaround a wrong infinite cost
When IntVF=1 (because VF was originally <vscale x 1>)
and ScalarC == VF.Cost and AssumedMinimumVscale == 1 (because
the min vscale is effectively 1 or it was not specified), then
we are going to compute MinTC1 as infinite. Because MinTC is
the std::ceil(std::max(MinTC1, MinTC2)) and std::ceil passes through
the infinite, the conversion to uint64_t triggers UB.

On x86, MinTC was 0 so we were accidentally triggering
a good path but on RISC-V MinTC was UINT64_MAX effectively
disabling vectorisation.

To mitigate this, make that AssumedMinimumVscale == 1 means
no information is available and we use a hard, simpler, threshold.