Fix to avoid crashing the compiler
We didn't realise that mask and the evl might not be dominating the gep
that we want to replace.

That said, though, as we only replace those cases where all the users
are vp.gather or vp.scatter with the same mask and evl it should be
possible to sink the new computation so it stays dominated by both the
mask and the evl.