Filter strided accesses
If the start (and perhaps the step?) has an SCEVAddRecExpr, the
LoopVectorizer has already modified the relevant IR and when we expand
it we're going to crash.

The actual solution is not to use SCEVExpander at all. In fact we should
have a SCEVVPlanExpander and represent the new IR we need in form of
VPlan recipes.

For now though, let's avoid the problematic cases by checking that the
SCEVAddRecExpr we've got is not at risk.