Improve matrix multiplication memory accesses.
Instead of creating a vector splat by loading multiple times the same
value from memory through the use of a vp_gather instruction, load the
value with a scalar load instruction and then splat it; in order to do
so, sinking the index operand of the vp_gather instruction to the same
basic block of the vp_gather itself is required.
As a bonus optimization, following `fmacc.vv` invocations recognize
the second operand to be a splatted value, falling back to the use of
`fmacc.vf/x` forms.