Fix DAG combiner
We were so focused on the offset being a constant (and the base pointer
too often being a zero) that we forgot the cases where this does not
hold.

Test MultiSource/Benchmarks/MiBench/consumer-lame has a file newmdct.c
where after inlining we end with a splat index but not a zero base
address. Handle this case as well.