[Driver] Add -mllvm= as an alias for -mllvm
Similar to D131455 (-Xclang=). As well as making it convenient for some build
systems, this allows `-Xarch_device '-mllvm=--inline-threshold=100'` (and
`-Xarch_host`; so we don't need to allow space separators which are uncommon in
driver code).

Bear in mind that -mllvm options are unstable and should be avoided if possible.

Reviewed By: jhuber6, yaxunl

Differential Revision: https://reviews.llvm.org/D143325