Disable in-loop reductions for now
This knob has the side-effect to enable in-loop reductions. In general
those are bad and they are better implemented outside of the loop but
have the advantage of not changing the order of the operations.

This change disables them for now until we use the proper vp.reductions
instead. This has been causing errors in our testsuite because the
emitted code is incorrect: we're reducing elements with undef values.